requirements.txt
opencv-python
pillow
numpy
svgwrite
scikit-image
